fix: add a hack to prevent lost entities from loading infinitely.
Not sure how it happened but some entities ended up
with their IDs referenced, but their data not pushed to the
syncserver. when you try to load these, if you don't have the
data locally, it will just load forever. This adds a hack that will
just create an empty entity with the ID after a 5 second timeout
whenever this happens.

Not a good solution, but we have changes we'll need to make
soon around this anyway, so it should smooth over most of the
issues temporarily.